+++ src/add-column-action-spacer/index.js
@@ -1,0 +1,1 @@
+export { default } from './add-column-action-spacer'